% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.seinfitR}
\alias{coef.seinfitR}
\title{Extract Coefficients}
\usage{
\method{coef}{seinfitR}(object, ...)
}
\arguments{
\item{object}{Object of class 'seinfitR'.}

\item{\dots}{currently unused.}
}
\value{
A named numeric vector containing the estimated model coefficients.
}
\description{
Extract model coefficients from a seinfitR object.
}
