% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasshouse.R
\docType{data}
\name{glasshouse}
\alias{glasshouse}
\title{Glasshouse Experiment Dataset}
\format{
A data frame with 14 rows and 2 columns:
\describe{
\item{p_i}{Nematode population density (initial population).}
\item{y}{Average crop yield at the given population density.}
}
}
\usage{
data(glasshouse, package = "seinfitR")
}
\description{
This dataset originates from a greenhouse experiment that assessed the effect of nematode population density on plant yield.
One cultivar was used, and 14 different nematode population densities (\code{p_i}), including zero, were tested.
Each density was replicated five times. The dataset provides the nematode densities and the corresponding average plant yield.
}
\references{
Schomaker, C., & Been, T. (2013). Plant growth and population dynamics. \emph{Plant Nematology}, 301-330. \doi{10.1079/9781780641515.0301}
}
\keyword{datasets}
