% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jambu}
\alias{jambu}
\title{colocar descrição}
\format{
A data frame with 1908 rows and four variables:
\describe{
\item{p_i}{xxx...}
\item{y}{xxx...}
}
}
\usage{
data(jambu, package = "seinfitR")
}
\description{
This dataset ...
}
\references{
xxxx
}
\keyword{datasets}
