% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seinfitR.R
\name{plot.seinfitR}
\alias{plot.seinfitR}
\title{Plot method for seinfitR objects}
\usage{
\method{plot}{seinfitR}(x, rel = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{seinfitR} (output from \code{seinfitR()}).}

\item{rel}{Logical. If \code{TRUE}, make a graph of the relative model.}

\item{\dots}{currently unused.}
}
\value{
A plot showing the observed data (blue points) and the fitted curve (red line).
}
\description{
This function generates a scatter plot of the observed data along with the fitted SEINH model.
}
