% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.seinfitR.R
\name{predict.seinfitR}
\alias{predict.seinfitR}
\title{Predict method for seinfitR objects}
\usage{
\method{predict}{seinfitR}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{seinfitR} (output from \code{seinfitR()}).}

\item{newdata}{Optional. A data frame containing the independent variable for which predictions should be made.
If not provided, predictions are made for the original data.}

\item{\dots}{currently unused.}
}
\value{
A data frame with the independent variable and the corresponding predicted values.
}
\description{
This function generates predictions based on a fitted Seinhorst model.
}
