% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.seinfitR}
\alias{print.seinfitR}
\title{Print seinfitR}
\usage{
\method{print}{seinfitR}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class 'seinfitR'.}

\item{digits}{minimal number of \emph{significant} digits, see
\code{\link[base]{print.default}}}

\item{\dots}{currently unused.}
}
\value{
No return value, called for side effects
}
\description{
Print contents of seinfitR object.
}
\seealso{
\code{\link{seinfitR}}
}
