% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{r_squared}
\alias{r_squared}
\title{R-squared Calculation}
\usage{
r_squared(object, ...)
}
\arguments{
\item{object}{Object of class 'seinfitR'.}

\item{\dots}{currently unused.}
}
\value{
A list with the following components:
\describe{
\item{R2}{The coefficient of determination (R-squared).}
\item{Adjusted_R2}{The adjusted R-squared value.}
}
}
\description{
Compute R-squared for seinfitR model.
}
