% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{seinfitR-methods}
\alias{seinfitR-methods}
\alias{print.seinfitR}
\alias{vcov.seinfitR}
\alias{summary.seinfitR}
\title{Methods for 'seinfitR' objects}
\usage{
\method{print}{seinfitR}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{vcov}{seinfitR}(object, ...)

\method{summary}{seinfitR}(object, ...)
}
\arguments{
\item{x}{an object of class \code{seinfitR}}

\item{digits}{minimal number of \emph{significant} digits, see
\code{\link[base]{print.default}}}

\item{\dots}{currently unused.}

\item{object}{An object of class \code{seinfitR} (output from \code{seinfitR()}).}
}
\value{
Prints a summary of the model fit to the console.
}
\description{
Methods for 'seinfitR' objects
}
