% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.seinfitR}
\alias{summary.seinfitR}
\title{Summary of seinfitR Model}
\usage{
\method{summary}{seinfitR}(object, ...)
}
\arguments{
\item{object}{Object of class 'seinfitR'.}

\item{\dots}{currently unused.}
}
\value{
No return value, called for side effects.
}
\description{
Display a summary of the seinfitR model.
}
\seealso{
\code{\link{seinfitR}}
}
