% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.seinfitR}
\alias{vcov.seinfitR}
\title{Variance-Covariance Matrix}
\usage{
\method{vcov}{seinfitR}(object, ...)
}
\arguments{
\item{object}{Object of class 'seinfitR'.}

\item{\dots}{currently unused.}
}
\value{
A matrix representing the covariance of the estimated coefficients.
}
\description{
Compute variance-covariance matrix for seinfitR model.
}
