% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jambu.R
\docType{data}
\name{jambu}
\alias{jambu}
\title{Jambu Dataset}
\format{
A data frame with 5,002 rows and 2 columns:
\describe{
\item{p_i}{Nematode population density (initial population)}
\item{y}{Crop yield, other plant growth parameter, or ratio of the estimated variable for plant growth in an initial population density of the nematode}
}
}
\source{
https://osf.io/pm94t
}
\usage{
data(jambu, package = "seinfitR")
}
\description{
This dataset contains information on nematode population density (\code{p_i}) and crop yield (\code{y}).
It is used in the \code{seinfitR} package for modeling plant growth in relation to nematode populations.
}
\references{
Silva, M.F., Faccioli, F.C., Honório, A.P. et al. (2024). First report of angular leaf spot in \emph{Acmella oleracea} caused by the foliar nematode \emph{Aphelenchoides pseudobesseyi}. \emph{J Plant Dis Prot}, 131, 1707–1720. https://doi.org/10.1007/s41348-024-00982-2
}
\keyword{datasets}
