% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seinfitR.R
\name{plot.seinfitR}
\alias{plot.seinfitR}
\title{Plot method for seinfitR objects}
\usage{
\method{plot}{seinfitR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{seinfitR} (output from \code{seinfitR()}).}

\item{...}{Additional arguments (not used).}
}
\value{
A plot showing the observed data (blue points) and the fitted curve (red line).
}
\description{
This function generates a scatter plot of the observed data along with the fitted SEINH model.
}
