% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seinfitR.R
\name{plot.seinfitR}
\alias{plot.seinfitR}
\title{Plot SeinfitR}
\usage{
\method{plot}{seinfitR}(x, rel = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{seinfitR} (output from \code{seinfitR()}).}

\item{rel}{Logical. If TRUE, the observed and fitted values are plotted relative to the maximum fitted value (normalized between 0 and 1). If FALSE, the original observed and fitted values are plotted.}

\item{\dots}{currently unused.}
}
\value{
A plot showing the observed data (blue points) and the fitted curve (red line).
}
\description{
Plot method for seinfitR objects
}
