% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seinfitR.R
\name{seinfitR}
\alias{seinfitR}
\title{Fit the Seinhorst Model to Experimental Data}
\usage{
seinfitR(data, start = NULL, control = seinfitR_control(maxiter = 100))
}
\arguments{
\item{data}{A data frame containing the experimental data. It should include at least
two columns: \code{x} (preplant nematode densities) and \code{y} (plant growth response).}

\item{start}{A list of starting values for the parameters \code{m}, \code{t}, and \code{z}. These values
are used to initialize the nonlinear least squares fitting process.}

\item{control}{control}
}
\value{
An object of class \code{nls}, which contains the results of the nonlinear least squares fitting,
including estimated parameters and residuals.
}
\description{
This function fits the Seinhorst equation to experimental data on the relationship between
preplant nematode densities and plant growth using nonlinear least squares fitting.
The function uses the \code{nlsLM} function from the \code{minpack.lm} package to estimate the parameters
of the model.
}
\examples{
library("readxl")
# Load example data from the inst/extdata folder
data <- read_excel(system.file("extdata", "raw_data_pathogenicity_test.xlsx", package = "seinfitR"))
colnames(data)[5] ="y"
colnames(data)[1] ="x"

# Fit the model using seinfitR with initial parameter values and control settings
seinfitR(data = data, start = list(m = 0.103, t = 250, z = 0.991),
         control = seinfitR_control(maxiter = 100))

}
